requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
curl_cffi==0.6.2
distro==1.9.0
flet==0.82.2
flet-desktop-light==0.82.2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
greenlet==3.0.3
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
msgpack==1.1.2
oauthlib==3.3.1
openai==1.13.3
packaging==26.0
playwright==1.42.0
pluggy==1.6.0
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.6.3
pydantic_core==2.16.3
pyee==11.0.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.0.1
repath==0.9.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
